requirements.txt
pyscreenshot
pillow
pynput
